+++ src/main.go
@@ -24,9 +24,10 @@ const (
 )
 
 var (
-	integrationVersion = "0.0.0"
-	gitCommit          = ""
-	buildDate          = ""
+	integrationVersion       = "0.0.0"
+	gitCommit                = ""
+	buildDate                = ""
+	errCheckingVarnishParams = errors.New("error occurred while checking for the existence of varnish.params at locations")
 )
 
 func main() {
@@ -127,25 +128,33 @@ func collectParamsFile(systemEntity *integration.Entity, argsParamLoc string) er
 // Debian/Ubuntu and RHEL/CentOS systems.
 func determineParamsFileLoc(argsParamLoc string) (*string, error) {
 	if argsParamLoc != "" {
-		if _, err := os.Stat(argsParamLoc); os.IsNotExist(err) {
+		if _, err := os.Stat(argsParamLoc); err != nil {
 			return nil, err
 		}
 		return &argsParamLoc, nil
 	}
 
+	var debianParamsLocErr, rhelParamsLocErr error
 	// Try Debian/Ubuntu path
 	paramsLoc := debianUbuntuParamsLoc
-	if _, err := os.Stat(paramsLoc); !os.IsNotExist(err) {
+	if _, debianParamsLocErr = os.Stat(paramsLoc); debianParamsLocErr == nil {
 		return &paramsLoc, nil
 	}
 
 	// Try RHEL/CentOS
 	paramsLoc = rhelCentosParamsLoc
-	if _, err := os.Stat(paramsLoc); !os.IsNotExist(err) {
+	if _, rhelParamsLocErr = os.Stat(paramsLoc); rhelParamsLocErr == nil {
 		return &paramsLoc, nil
 	}
 
-	return nil, errors.New("varnish.params file could not be found")
+	// Debug log to print the reasons for errors
+	log.Debug("%s: '%s' - %s, '%s' - %s",
+		errCheckingVarnishParams,
+		debianUbuntuParamsLoc, debianParamsLocErr.Error(),
+		rhelCentosParamsLoc, rhelParamsLocErr.Error())
+
+	return nil, fmt.Errorf("%w: '%s' and '%s'",
+		errCheckingVarnishParams, debianUbuntuParamsLoc, rhelCentosParamsLoc)
 }
 
 func parseParamsFile(file *os.File) (map[string]string, error) {