+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Unreleased section should follow [Release Toolkit](https://github.com/newrelic/r
 
 ## Unreleased
 
+### bugfix
+- Fixed error management while checking existence of varnish.params file
+
 ## v2.6.0 - 2024-10-14
 
 ### dependency